scripts/demo/requirements-live-dashboards.txt
# Live Dashboard Dependencies
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
prometheus-client>=0.15.0
mlflow

# Optional: For Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0
